TripodMotionControl: only one speed reference for all joints
Allow setting only one value of speed for all joints; getting speeds will
always return the same reference value for all of them.

NOTE: this value will be the fastest speed for the movement, when more
joints move at the same time, the joints with shorter movement to do
will move slower to be in synch. (start and end the movement at the
same time)