tripodMotionControl: fix set position limits for angular joints
Although this device is not meant to scale with number of joint (a tripod has three
joints) a safety check has been added with respect to joint numbber.

Changing the config file this could became scalable as well. Ratio: this could in
future became a general wrapper for iKin library